+++ core/components/romanesco/elements/snippets/06_formulas/f_performance/minifycss.snippet.php
@@ -57,7 +57,7 @@ $pendingTasks = $modx->getCollection('sTaskRun', [
 ]);
 foreach ($pendingTasks as $pendingTask) {
     $data = $pendingTask->get('data');
-    if ($data['css_path'] ?? '' == $cssPath) {
+    if (isset($data['css_path']) && $data['css_path'] == $cssPath) {
         return;
     }
 }